+++ haskell-project/src/Parse.hs
@@ -30,7 +30,7 @@ data Location = Location
 
 data Current = Current
   { observation_time :: String
-  , temperature :: Int
+  , temperature :: Float
   , weather_code :: Int
   , weather_icons :: [String]
   , weather_descriptions :: [String]
@@ -38,7 +38,7 @@ data Current = Current
   , wind_degree :: Int
   , wind_dir :: String
   , pressure :: Int
-  , precip :: Int
+  , precip :: Float
   , humidity :: Int
   , cloudcover :: Int
   , feelslike :: Int
@@ -65,12 +65,12 @@ instance FromJSON Current
 instance FromJSON WeatherData
 
 -- Determine rain status based on precipitation value
-getRainStatus :: Int -> String
+getRainStatus :: Float -> String
 getRainStatus precipValue
     | precipValue < 50 = "no"
     | otherwise = "yes"
 
-parseWeatherData :: String -> IO (String,Int, Int, String)
+parseWeatherData :: String -> IO (String, Float, Float, String)
 parseWeatherData filePath = do
     -- Read the JSON file
     jsonContents <- B.readFile filePath
@@ -91,7 +91,7 @@ parseWeatherData filePath = do
                 rainStatus = getRainStatus preci
             return (place, temp, preci, rainStatus)
 
-parse :: IO (String, Int, String)
+parse :: IO (String, Float, String)
 parse = do
     (place, temp, preci, rainStatus) <- parseWeatherData "data/response.json"
     return (place,temp,rainStatus)
